refactor(schlib): extract read_io.rs and write_io.rs from mod.rs
Pure code motion (U2 of the SchLib/PcbLib structural unification): the
OLE stream orchestration moves out of mod.rs into read_io.rs and
write_io.rs, mirroring pcblib's split. mod.rs keeps the SchLib/Symbol
types, CRUD and tests. No file outside src/altium/schlib/ changed; no
behaviour change; gates green incl. the readability oracle
(0 regressions).
